Split files to smaller ones and tidy up the whole test package
