+++ memex/src/test/java/com/johnlpage/memex/cucumber/CucumberTestsContainersConfig.java
@@ -1,6 +1,6 @@
 package com.johnlpage.memex.cucumber;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Condition;
@@ -10,13 +10,6 @@ import org.springframework.core.type.AnnotatedTypeMetadata;
 import org.springframework.test.context.DynamicPropertyRegistrar;
 import org.testcontainers.mongodb.MongoDBAtlasLocalContainer;
 
-class MongoUriMissingCondition implements Condition {
-    @Override
-    public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
-        String uri = context.getEnvironment().getProperty("spring.data.mongodb.uri");
-        return uri == null || uri.trim().isEmpty();
-    }
-}
 
 @TestConfiguration
 public class CucumberTestsContainersConfig {
@@ -35,5 +28,13 @@ public class CucumberTestsContainersConfig {
             registry.add("spring.data.mongodb.uri", mongoDBContainer::getConnectionString);
             registry.add("spring.data.mongodb.database", () -> "memex");
         };
+    }
+
+    static class MongoUriMissingCondition implements Condition {
+        @Override
+        public boolean matches(ConditionContext context, @NotNull AnnotatedTypeMetadata metadata) {
+            String uri = context.getEnvironment().getProperty("spring.data.mongodb.uri");
+            return uri == null || uri.trim().isEmpty();
+        }
     }
 }
