+++ memex/src/test/resources/application-test.properties
@@ -1,8 +1,7 @@
 #Keep this empty if you want tests to use Mongo test container
 spring.data.mongodb.uri=
 spring.data.mongodb.database=memex
+memex.kafkaexmple.enabled=true
 
 memex.test.data.vehicleinspection-testid-range.start=10000
-memex.test.data.vehicleinspection-testid-range.end=20000
-
-memex.kafkaexmple.enabled=true
+memex.test.data.vehicleinspection-testid-range.end=11000
