+++ memex/src/test/resources/features/inspections.kafka.feature
@@ -1,5 +1,10 @@
-Feature: Vehicle Inspection Kafka consumer
+@kafka @vehicle_inspection
+Feature: Vehicle Inspection Kafka Integration
+  This feature tests the integration of vehicle inspections with Kafka.
+  It ensures that messages sent to Kafka topics are correctly consumed and processed,
+  leading to the appropriate storage or update of inspection data in the system.
 
+  @kafka @sunny_day
   Scenario: Vehicle Inspection Kafka consumer listens to sent messages
     Given the vehicle inspections in range 10000-11000 do not exist
     When I send 100 vehicle inspections starting with id 10000 to kafka with:
@@ -11,4 +16,3 @@ Feature: Vehicle Inspection Kafka consumer
     """
     {"capacity": 60, "vehicle": {"make": "Ford"}}
     """
-
